Add audio utilities, use them where possible.
