+++ mlx_audio/stt/models/parakeet/audio.py
@@ -1,15 +1,8 @@
-import functools
-import math
-import shutil
 from dataclasses import dataclass
-from functools import lru_cache
-from pathlib import Path
-from subprocess import CalledProcessError, run
-from typing import Optional, Union
 
-import librosa
 import mlx.core as mx
-import numpy as np
+
+from mlx_audio.utils import hanning, hamming, blackman, bartlett, mel_filters, stft, STR_TO_WINDOW_FN
 
 
 @dataclass
@@ -33,151 +26,6 @@ class PreprocessArgs:
     def hop_length(self) -> int:
         return int(self.window_stride * self.sample_rate)
 
-
-@lru_cache(maxsize=None)
-def hanning(size):
-    return mx.array(
-        [0.5 * (1 - math.cos(2 * math.pi * n / (size - 1))) for n in range(size)]
-    )
-
-
-@lru_cache(maxsize=None)
-def hamming(size):
-    return mx.array(
-        [0.54 - 0.46 * math.cos(2 * math.pi * n / (size - 1)) for n in range(size)]
-    )
-
-
-@lru_cache(maxsize=None)
-def blackman(size):
-    return mx.array(
-        [
-            0.42
-            - 0.5 * math.cos(2 * math.pi * n / (size - 1))
-            + 0.08 * math.cos(4 * math.pi * n / (size - 1))
-            for n in range(size)
-        ]
-    )
-
-
-@lru_cache(maxsize=None)
-def bartlett(size):
-    return mx.array([1 - 2 * abs(n - (size - 1) / 2) / (size - 1) for n in range(size)])
-
-
-def stft(
-    x, n_fft, hop_length=None, win_length=None, window=None, axis=-1, pad_mode="reflect"
-):
-    if win_length is None:
-        win_length = n_fft
-    if hop_length is None:
-        hop_length = n_fft // 4
-    if window is None:
-        window = mx.ones(win_length)
-
-    if win_length != n_fft:
-        if win_length > n_fft:
-            window = window[:n_fft]
-        else:
-            padding = [(0, n_fft - win_length)]
-            window = mx.pad(window, padding)
-
-    def _pad(x, padding, pad_mode="constant"):
-        if pad_mode == "constant":
-            return mx.pad(x, [(padding, padding)])
-        elif pad_mode == "reflect":
-            prefix = x[1 : padding + 1][::-1]
-            suffix = x[-(padding + 1) : -1][::-1]
-            return mx.concatenate([prefix, x, suffix])
-        else:
-            raise ValueError(f"Invalid pad_mode {pad_mode}")
-
-    padding = win_length // 2
-    x = _pad(x, padding, pad_mode)
-
-    strides = [hop_length, 1]
-    t = (x.size - win_length + hop_length) // hop_length
-    shape = [t, n_fft]
-    x = mx.as_strided(x, shape=shape, strides=strides)
-    return mx.fft.rfft(x * window)
-
-
-@lru_cache(maxsize=None)
-def mel_filters(
-    sample_rate: int,
-    n_fft: int,
-    n_mels: int,
-    f_min: float = 0,
-    f_max: Optional[float] = None,
-    norm: Optional[str] = None,
-    mel_scale: Optional[str] = "htk",
-) -> mx.array:
-    def hz_to_mel(freq, mel_scale="htk"):
-        if mel_scale == "htk":
-            return 2595.0 * math.log10(1.0 + freq / 700.0)
-
-        # slaney scale
-        f_min, f_sp = 0.0, 200.0 / 3
-        mels = (freq - f_min) / f_sp
-        min_log_hz = 1000.0
-        min_log_mel = (min_log_hz - f_min) / f_sp
-        logstep = math.log(6.4) / 27.0
-        if freq >= min_log_hz:
-            mels = min_log_mel + math.log(freq / min_log_hz) / logstep
-        return mels
-
-    def mel_to_hz(mels, mel_scale="htk"):
-        if mel_scale == "htk":
-            return 700.0 * (10.0 ** (mels / 2595.0) - 1.0)
-
-        # slaney scale
-        f_min, f_sp = 0.0, 200.0 / 3
-        freqs = f_min + f_sp * mels
-        min_log_hz = 1000.0
-        min_log_mel = (min_log_hz - f_min) / f_sp
-        logstep = math.log(6.4) / 27.0
-        freqs = mx.where(
-            mels >= min_log_mel,
-            min_log_hz * mx.exp(logstep * (mels - min_log_mel)),
-            freqs,
-        )
-        return freqs
-
-    f_max = f_max or sample_rate / 2
-
-    # generate frequency points
-
-    n_freqs = n_fft // 2 + 1
-    all_freqs = mx.linspace(0, sample_rate // 2, n_freqs)
-
-    # convert frequencies to mel and back to hz
-
-    m_min = hz_to_mel(f_min, mel_scale)
-    m_max = hz_to_mel(f_max, mel_scale)
-    m_pts = mx.linspace(m_min, m_max, n_mels + 2)
-    f_pts = mel_to_hz(m_pts, mel_scale)
-
-    # compute slopes for filterbank
-
-    f_diff = f_pts[1:] - f_pts[:-1]
-    slopes = mx.expand_dims(f_pts, 0) - mx.expand_dims(all_freqs, 1)
-
-    # calculate overlapping triangular filters
-
-    down_slopes = (-slopes[:, :-2]) / f_diff[:-1]
-    up_slopes = slopes[:, 2:] / f_diff[1:]
-    filterbank = mx.maximum(
-        mx.zeros_like(down_slopes), mx.minimum(down_slopes, up_slopes)
-    )
-
-    if norm == "slaney":
-        enorm = 2.0 / (f_pts[2 : n_mels + 2] - f_pts[:n_mels])
-        filterbank *= mx.expand_dims(enorm, 0)
-
-    filterbank = filterbank.moveaxis(0, 1)
-    return filterbank
-
-
 def log_mel_spectrogram(x: mx.array, args: PreprocessArgs) -> mx.array:
     original_dtype = x.dtype
 
@@ -186,23 +34,8 @@ def log_mel_spectrogram(x: mx.array, args: PreprocessArgs) -> mx.array:
             pad_length = args.pad_to - x.shape[-1]
             x = mx.pad(x, ((0, pad_length),), constant_values=args.pad_value)
 
-    window = (
-        hanning(args.win_length).astype(x.dtype)
-        if args.window == "hanning"
-        else (
-            hamming(args.win_length).astype(x.dtype)
-            if args.window == "hamming"
-            else (
-                blackman(args.win_length).astype(x.dtype)
-                if args.window == "blackman"
-                else (
-                    bartlett(args.win_length).astype(x.dtype)
-                    if args.window == "bartlett"
-                    else None
-                )
-            )
-        )
-    )
+    window_fn = STR_TO_WINDOW_FN.get(args.window, None)
+    window = window_fn(args.win_length) if window_fn else hanning(args.win_length)
 
     x = stft(x, args.n_fft, args.hop_length, args.win_length, window)
     x = mx.square(mx.abs(x)).astype(original_dtype)
