+++ mlx_audio/stt/models/whisper/audio.py
@@ -1,14 +1,12 @@
 # Copyright © 2023 Apple Inc.
 
-import math
-from functools import lru_cache
-from typing import Optional, Union
+from typing import Union
 
 import mlx.core as mx
 import numpy as np
-from scipy import signal
 
 from mlx_audio.stt.utils import load_audio
+from mlx_audio.utils import hanning, mel_filters, stft
 
 # hard-coded audio hyperparameters
 SAMPLE_RATE = 16000
@@ -39,114 +37,6 @@ def pad_or_trim(array, length: int = N_SAMPLES, *, axis: int = -1):
 
     return array
 
-
-@lru_cache(maxsize=None)
-def mel_filters(
-    sample_rate: int,
-    n_fft: int,
-    n_mels: int,
-    f_min: float = 0,
-    f_max: Optional[float] = None,
-    norm: Optional[str] = None,
-    mel_scale: Optional[str] = "htk",
-) -> mx.array:
-    def hz_to_mel(freq, mel_scale="htk"):
-        if mel_scale == "htk":
-            return 2595.0 * math.log10(1.0 + freq / 700.0)
-
-        # slaney scale
-        f_min, f_sp = 0.0, 200.0 / 3
-        mels = (freq - f_min) / f_sp
-        min_log_hz = 1000.0
-        min_log_mel = (min_log_hz - f_min) / f_sp
-        logstep = math.log(6.4) / 27.0
-        if freq >= min_log_hz:
-            mels = min_log_mel + math.log(freq / min_log_hz) / logstep
-        return mels
-
-    def mel_to_hz(mels, mel_scale="htk"):
-        if mel_scale == "htk":
-            return 700.0 * (10.0 ** (mels / 2595.0) - 1.0)
-
-        # slaney scale
-        f_min, f_sp = 0.0, 200.0 / 3
-        freqs = f_min + f_sp * mels
-        min_log_hz = 1000.0
-        min_log_mel = (min_log_hz - f_min) / f_sp
-        logstep = math.log(6.4) / 27.0
-        freqs = mx.where(
-            mels >= min_log_mel,
-            min_log_hz * mx.exp(logstep * (mels - min_log_mel)),
-            freqs,
-        )
-        return freqs
-
-    f_max = f_max or sample_rate / 2
-
-    # generate frequency points
-
-    n_freqs = n_fft // 2 + 1
-    all_freqs = mx.linspace(0, sample_rate // 2, n_freqs)
-
-    # convert frequencies to mel and back to hz
-
-    m_min = hz_to_mel(f_min, mel_scale)
-    m_max = hz_to_mel(f_max, mel_scale)
-    m_pts = mx.linspace(m_min, m_max, n_mels + 2)
-    f_pts = mel_to_hz(m_pts, mel_scale)
-
-    # compute slopes for filterbank
-
-    f_diff = f_pts[1:] - f_pts[:-1]
-    slopes = mx.expand_dims(f_pts, 0) - mx.expand_dims(all_freqs, 1)
-
-    # calculate overlapping triangular filters
-
-    down_slopes = (-slopes[:, :-2]) / f_diff[:-1]
-    up_slopes = slopes[:, 2:] / f_diff[1:]
-    filterbank = mx.maximum(
-        mx.zeros_like(down_slopes), mx.minimum(down_slopes, up_slopes)
-    )
-
-    if norm == "slaney":
-        enorm = 2.0 / (f_pts[2 : n_mels + 2] - f_pts[:n_mels])
-        filterbank *= mx.expand_dims(enorm, 0)
-
-    filterbank = filterbank.moveaxis(0, 1)
-    return filterbank
-
-
-@lru_cache(maxsize=None)
-def hanning(size):
-    return mx.array(
-        [0.5 * (1 - math.cos(2 * math.pi * n / (size - 1))) for n in range(size)]
-    )
-
-
-def stft(x, window, nperseg=256, noverlap=None, nfft=None, axis=-1, pad_mode="reflect"):
-    if nfft is None:
-        nfft = nperseg
-    if noverlap is None:
-        noverlap = nfft // 4
-
-    def _pad(x, padding, pad_mode="constant"):
-        if pad_mode == "constant":
-            return mx.pad(x, [(padding, padding)])
-        elif pad_mode == "reflect":
-            prefix = x[1 : padding + 1][::-1]
-            suffix = x[-(padding + 1) : -1][::-1]
-            return mx.concatenate([prefix, x, suffix])
-        else:
-            raise ValueError(f"Invalid pad_mode {pad_mode}")
-
-    padding = nperseg // 2
-    x = _pad(x, padding, pad_mode)
-
-    strides = [noverlap, 1]
-    t = (x.size - nperseg + noverlap) // noverlap
-    shape = [t, nfft]
-    x = mx.as_strided(x, shape=shape, strides=strides)
-    return mx.fft.rfft(x * window)
 
 
 def log_mel_spectrogram(
@@ -181,7 +71,7 @@ def log_mel_spectrogram(
     if padding > 0:
         audio = mx.pad(audio, (0, padding))
     window = hanning(N_FFT)
-    freqs = stft(audio, window, nperseg=N_FFT, noverlap=HOP_LENGTH)
+    freqs = stft(audio, window=window, n_fft=N_FFT, hop_length=HOP_LENGTH)
     magnitudes = freqs[:-1, :].abs().square()
 
     filters = mel_filters(SAMPLE_RATE, N_FFT, n_mels, norm="slaney", mel_scale=None)
