+++ mlx_audio/tts/models/spark/bicodec.py
@@ -7,7 +7,7 @@ import numpy as np
 from omegaconf import DictConfig
 from safetensors.torch import load_file
 
-from mlx_audio.codec.models.vocos.mel import hanning, mel_filters, stft
+from mlx_audio.utils import hanning, mel_filters, stft
 from mlx_audio.tts.models.spark.modules.encoder_decoder.feat_decoder import Decoder
 from mlx_audio.tts.models.spark.modules.encoder_decoder.feat_encoder import Encoder
 from mlx_audio.tts.models.spark.modules.encoder_decoder.wave_generator import (
